ref(bot): move check run tests into separate file
